refactor: un-Class-ify the `Teacher` model & separate scraper's logic out!
